Create The Rectangle class
Design a class named Rectangle to represent a rectangle. The class contains:

Two double data fields named width and height that specify the width and height of the rectangle.

A no-arg constructor that creates a rectangle with width 1 and height 1.

A constructor that creates a default rectangle with the specified width and height.

The accessor and mutator functions for all the data fields.

A function named getArea() that returns the area of this rectangle.

A function named getPerimeter() that returns the perimeter.